+++ cadwyn/_utils.py
@@ -1,6 +1,8 @@
 from collections.abc import Callable
 from typing import Any, Generic, TypeVar, Union
 
+from pydantic._internal._decorators import unwrap_wrapped_function
+
 Sentinel: Any = object()
 UnionType = type(int | str) | type(Union[int, str])
 _T = TypeVar("_T", bound=Callable)
@@ -31,3 +33,10 @@ def same_definition_as_in(t: _T) -> Callable[[Callable], _T]:
         return f  # pyright: ignore[reportReturnType]
 
     return decorator
+
+
+def _fully_unwrap_decorator(func: Callable, is_pydantic_v1_style_validator: Any):
+    func = unwrap_wrapped_function(func)
+    if is_pydantic_v1_style_validator and func.__closure__:
+        func = func.__closure__[0].cell_contents
+    return func